Image + Texture asset wip

+++ Phezu/Source/Asset/Types/ImageAsset.cpp
@@ -7,7 +7,7 @@ namespace Phezu {
 
 	void ImageAsset::Deserialize(const std::string& data) {
 		int channelsInFile = 0;
-		Data = stbi_load_from_memory(reinterpret_cast<const stbi_uc*>(data.data()), data.size(), &Width, &Height, &channelsInFile, 4);
+		Data = stbi_load_from_memory(reinterpret_cast<const stbi_uc*>(data.data()), static_cast<int>(data.size()), &Width, &Height, &channelsInFile, 4);
 	}
 
 }
